install bundle tool set one by one
Due to network problem, it can't install many bundle tool sets at one
time successfully. So, it's better to install bundle tool set one by one
and use the cached fs layer to build the docker image.
This is not a goole idea since it'll extend the fs layer and make bigger
image. However, it's useful in bad network nevironment and useful for
debug purpose

Singed-off-by: Liu Changcheng <changcheng.liu@aliyun.com>